+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unpublished
+## [1.3.0] - 2020-05-07
 
 ### Added
 
@@ -10,6 +10,8 @@
 
 - Only pass down unused props to the wrapped component
 - Added `ViewProps` to Block tipes
+- Renamed `withBlock` to `asBlock` (`withBlock` still works, but is deprecated)
+- `asBlock` is now forwarding `ref` to the original component
 
 ## [1.2.0] - 2019-12-13